Assigment for lesson 1.2
