+++ src/App.js
@@ -1,29 +1,13 @@
 import React from 'react';
-
-const todoList = [
-    {
-        id: 1,
-        title: 'Create Empty Array',
-    },
-    {
-        id: 2,
-        title: 'Create 3 objects',
-    },
-    {
-        id: 3,
-        title: 'Complete assignment',
-    },
-];
+import TodoList from './TodoList';
+import AddTodoForm from './AddTodoForm';
 
 function App() {
     return (
         <div>
             <h1>Todo List</h1>
-            <ul>
-                {todoList.map(function (item) {
-                    return <li key={item.id}>{item.title}</li>;
-                })}
-            </ul>
+            <AddTodoForm />
+            <TodoList />
         </div>
     );
 }
